Refactor wager updater to its own class
Probably technically reusable per game manager, too.
